1、更新t_sys_log表，去除user_id字段，改成user_account字段，用于记录操作的执行人 2、SysLog注解移动到zywork-log-aspect模块中 3、SysLogAspect类中使用Shiro的subject记录操作的执行人 4、更新FileUtils工具类 5、更新view_search_js模板和ViewGenerator类

+++ zywork-common/src/main/java/top/zywork/common/FileUtils.java
@@ -161,13 +161,15 @@ public class FileUtils {
      */
     public static void deleteFiles(String srcDir) {
         File file = new File(srcDir);
-        File[] files = file.listFiles();
-        if (files != null && files.length > 0) {
-            for (File f : files) {
-                if (f.isDirectory()) {
-                    deleteFiles(f.getAbsolutePath());
-                } else {
-                    f.delete();
+        if (file.isDirectory()) {
+            File[] files = file.listFiles();
+            if (files != null && files.length > 0) {
+                for (File f : files) {
+                    if (f.isDirectory()) {
+                        deleteFiles(f.getAbsolutePath());
+                    } else {
+                        f.delete();
+                    }
                 }
             }
         }
